Unified TrackedInpoutContextSelectMulti into TrackedInputContext
